move 'Reset Settings' button to settings sidebar
